+++ CHANGELOG.md
@@ -21,3 +21,4 @@
 - Fixed Frog Blessing Display not working
 - Fixed Night Squid tracking and achievements (renamed to Inkling in-game)
 - Fixed rfupf member count not properly updating
+- Fixed Supporting the cause achievement :dog: